Re-added wires, and changed the organisational scheme a little (all content, e.g. the Object, Cell, Part, Renderer go into the same file)
