+++ src/main/java/org/eln2/mc/common/parts/PartRegistry.kt
@@ -5,6 +5,7 @@ import net.minecraft.world.phys.Vec3
 import net.minecraftforge.eventbus.api.IEventBus
 import net.minecraftforge.registries.*
 import org.eln2.mc.Eln2
+import org.eln2.mc.common.content.WirePart
 import org.eln2.mc.common.items.foundation.PartItem
 import org.eln2.mc.common.parts.foundation.PartProvider
 import org.eln2.mc.common.parts.foundation.providers.BasicPartProvider
@@ -55,7 +56,7 @@ object PartRegistry {
         return ForgeRegistries.ITEMS.getValue(id) as PartItem
     }
 
-    //val WIRE_PART = part("wire_part", BasicPartProvider(::WirePart, Vec3(0.1, 0.1, 0.1)))
+    val WIRE_PART = part("wire_part", BasicPartProvider(::WirePart, Vec3(0.1, 0.1, 0.1)))
     //val BATTERY_PART = part("battery_part", BasicPartProvider(::BatteryPart, Vec3(0.8, 0.8, 0.8)))
     val RESISTOR_PART = part("resistor_part", BasicPartProvider(::ResistorPart, Vec3(1.0, 1.0, 1.0)))
     val VOLTAGE_SOURCE_PART = part("voltage_source_part", BasicPartProvider(::VoltageSourcePart, Vec3(1.0, 1.0, 1.0)))
